Fix folder structure, rename eeprom
The folder structure was changed as to not feature more than 2 levels of
intendation.

Also building on Windows hosts was fixed. Building fails on Windows
because Windows has a case-insensitive filesystem but the code had a
"eeprom.h" and a framework-provided "EEPROM.h" -- these two clash.
