+++ .settings/org.eclipse.cdt.core.prefs
@@ -1,0 +1,11 @@
+eclipse.preferences.version=1
+environment/project/0.910961921/PATH/delimiter=;
+environment/project/0.910961921/PATH/operation=replace
+environment/project/0.910961921/PATH/value=C\:\\Program Files (x86)\\STMicroelectronics\\st_toolset\\asm;C\:\\Program Files (x86)\\NVIDIA Corporation\\PhysX\\Common;C\:\\ProgramData\\Oracle\\Java\\javapath;C\:\\WINDOWS\\system32;C\:\\WINDOWS;C\:\\WINDOWS\\System32\\Wbem;C\:\\WINDOWS\\System32\\WindowsPowerShell\\v1.0\\;C\:\\Program Files (x86)\\GNU\\GnuPG\\pub;C\:\\Program Files (x86)\\mingw-w64\\i686-7.2.0-posix-dwarf-rt_v5-rev1\\mingw32\\bin;C\:\\Users\\Maxi\\Downloads\\GetGnuWin32\\gnuwin32\\bin;C\:\\Program Files\\Microsoft SQL Server\\120\\Tools\\Binn\\;C\:\\Program Files\\Microsoft SQL Server\\130\\Tools\\Binn\\;C\:\\Program Files (x86)\\GtkSharp\\2.12\\bin;C\:\\Users\\Maxi\\Downloads\\nano-1.0.8;C\:\\Program Files\\nodejs\\;C\:\\Program Files\\MiKTeX 2.9\\miktex\\bin\\x64\\;C\:\\Python27\\;C\:\\Python27\\Scripts;C\:\\Program Files (x86)\\MSBuild\\14.0\\Bin;C\:\\Users\\Maxi\\nasm-2.12.02;C\:\\Users\\Maxi\\Downloads\\fasmw17158\\;C\:\\Windows\\twain_32\\CNQ4200;C\:\\Program Files (x86)\\Python35\\Scripts;C\:\\Users\\Maxi\\Downloads\\ffmpeg-20161230-6993bb4-win64-shared\\bin;C\:\\Users\\Maxi\\Desktop\\esp-idf\\components\\partition_table;C\:\\Users\\Maxi\\Desktop\\esp-idf\\components\\esptool_py\\esptool;C\:\\Program Files\\Java\\jdk1.8.0_111\\bin;C\:\\Program Files\\OpenVPN\\bin;C\:\\Program Files\\TortoiseGit\\bin;C\:\\Program Files (x86)\\Skype\\Phone\\;C\:\\Users\\Maxi\\Downloads\\arduino-1.8.1\\hardware\\tools\\avr\\bin;C\:\\Users\\Maxi\\Desktop\\Programming stuff\\golink\\Golink;C\:\\Users\\Maxi\\Desktop\\STM8 Dev\\sdcc\\bin;C\:\\Program Files (x86)\\COSMIC\\FSE_Compilers\\CXSTM8;C\:\\Users\\Maxi\\Downloads\\esp32_win32_msys2_environment_and_toolchain-20170330\\msys32\\opt\\xtensa-esp32-elf\\bin;C\:\\Program Files (x86)\\Windows Live\\Shared;C\:\\Program Files\\dotnet\\;C\:\\Users\\Maxi\\Desktop\\Programming_stuff\\STM8_Dev\\sdcc\\bin;C\:\\Users\\Maxi\\Downloads\\esp32_win32_msys2_environment_and_toolchain-20170330\\msys32\\opt\\openocd-esp32\\bin;C\:\\Users\\Maxi\\Desktop\\GHDL\\v0.35-mingw64-llvm\\bin;C\:\\Program Files\\Git\\cmd;C\:\\WINDOWS\\System32\\OpenSSH\\;C\:\\Users\\Maxi\\AppData\\Local\\Microsoft\\WindowsApps;C\:\\Program Files (x86)\\Nmap;C\:\\Users\\Maxi\\AppData\\Roaming\\npm;C\:\\Users\\Maxi\\AppData\\Local\\Apps\\cURL\\bin;C\:\\Program Files (x86)\\COSMIC\\FSE_Compilers\\CXSTM8;C\:\\Program Files (x86)\\COSMIC\\FSE_Compilers\\CXSTM32;
+environment/project/0.910961921/append=true
+environment/project/0.910961921/appendContributed=true
+environment/project/0.910961921.1363900502/PATH/delimiter=;
+environment/project/0.910961921.1363900502/PATH/operation=replace
+environment/project/0.910961921.1363900502/PATH/value=C\:\\Program Files (x86)\\STMicroelectronics\\st_toolset\\asm;C\:\\Program Files (x86)\\NVIDIA Corporation\\PhysX\\Common;C\:\\ProgramData\\Oracle\\Java\\javapath;C\:\\WINDOWS\\system32;C\:\\WINDOWS;C\:\\WINDOWS\\System32\\Wbem;C\:\\WINDOWS\\System32\\WindowsPowerShell\\v1.0\\;C\:\\Program Files (x86)\\GNU\\GnuPG\\pub;C\:\\Program Files (x86)\\mingw-w64\\i686-7.2.0-posix-dwarf-rt_v5-rev1\\mingw32\\bin;C\:\\Users\\Maxi\\Downloads\\GetGnuWin32\\gnuwin32\\bin;C\:\\Program Files\\Microsoft SQL Server\\120\\Tools\\Binn\\;C\:\\Program Files\\Microsoft SQL Server\\130\\Tools\\Binn\\;C\:\\Program Files (x86)\\GtkSharp\\2.12\\bin;C\:\\Users\\Maxi\\Downloads\\nano-1.0.8;C\:\\Program Files\\nodejs\\;C\:\\Program Files\\MiKTeX 2.9\\miktex\\bin\\x64\\;C\:\\Python27\\;C\:\\Python27\\Scripts;C\:\\Program Files (x86)\\MSBuild\\14.0\\Bin;C\:\\Users\\Maxi\\nasm-2.12.02;C\:\\Users\\Maxi\\Downloads\\fasmw17158\\;C\:\\Windows\\twain_32\\CNQ4200;C\:\\Program Files (x86)\\Python35\\Scripts;C\:\\Users\\Maxi\\Downloads\\ffmpeg-20161230-6993bb4-win64-shared\\bin;C\:\\Users\\Maxi\\Desktop\\esp-idf\\components\\partition_table;C\:\\Users\\Maxi\\Desktop\\esp-idf\\components\\esptool_py\\esptool;C\:\\Program Files\\Java\\jdk1.8.0_111\\bin;C\:\\Program Files\\OpenVPN\\bin;C\:\\Program Files\\TortoiseGit\\bin;C\:\\Program Files (x86)\\Skype\\Phone\\;C\:\\Users\\Maxi\\Downloads\\arduino-1.8.1\\hardware\\tools\\avr\\bin;C\:\\Users\\Maxi\\Desktop\\Programming stuff\\golink\\Golink;C\:\\Users\\Maxi\\Desktop\\STM8 Dev\\sdcc\\bin;C\:\\Program Files (x86)\\COSMIC\\FSE_Compilers\\CXSTM8;C\:\\Users\\Maxi\\Downloads\\esp32_win32_msys2_environment_and_toolchain-20170330\\msys32\\opt\\xtensa-esp32-elf\\bin;C\:\\Program Files (x86)\\Windows Live\\Shared;C\:\\Program Files\\dotnet\\;C\:\\Users\\Maxi\\Desktop\\Programming_stuff\\STM8_Dev\\sdcc\\bin;C\:\\Users\\Maxi\\Downloads\\esp32_win32_msys2_environment_and_toolchain-20170330\\msys32\\opt\\openocd-esp32\\bin;C\:\\Users\\Maxi\\Desktop\\GHDL\\v0.35-mingw64-llvm\\bin;C\:\\Program Files\\Git\\cmd;C\:\\WINDOWS\\System32\\OpenSSH\\;C\:\\Users\\Maxi\\AppData\\Local\\Microsoft\\WindowsApps;C\:\\Program Files (x86)\\Nmap;C\:\\Users\\Maxi\\AppData\\Roaming\\npm;C\:\\Users\\Maxi\\AppData\\Local\\Apps\\cURL\\bin;C\:\\Program Files (x86)\\COSMIC\\FSE_Compilers\\CXSTM8;C\:\\Program Files (x86)\\COSMIC\\FSE_Compilers\\CXSTM32;
+environment/project/0.910961921.1363900502/append=true
+environment/project/0.910961921.1363900502/appendContributed=true
