+++ lib/automation/automation.cpp
