+++ lib/automation/automation.h
