+++ lib/class_commons/class_commons.cpp
