+++ lib/class_commons/class_commons.h
