+++ lib/scenario/scenario.cpp
