+++ lib/ser-deser/common_deser.cpp
