+++ platformio.ini
@@ -12,7 +12,11 @@
 platform = espressif8266
 board = nodemcuv2
 framework = arduino
-lib_ldf_mode = chain+
+;lib_ldf_mode = chain+
+lib_deps = 
+	ArduinoJson
+	IRremoteESP8266
+
 
 ;[env:esp8266]
 ;platform = espressif8266
